+++ R/eta_squared_posterior.R
@@ -146,11 +146,13 @@ eta_squared_posterior.brmsfit <- eta_squared_posterior.stanreg
                                 function(xi) isTRUE(all.equal(mean(xi),0)))
   }
 
-  if (length(factors)) {
+
+  of <- options()$contrasts
+  if (length(factors) && !all(of %in% c('contr.sum', 'contr.poly', "contr.bayes", "contr.helmert"))) {
     # if a contrast has negative and positive values, it is assumed to be one of:
     # "contr.sum", "contr.helmert", "contr.poly", "contr.bayes"
     factors_centered <- sapply(X[, factors, drop = FALSE],
-                               function (xi) any(contrasts(xi) < 0) & any(contrasts(xi) > 0))
+                               function(xi) any(contrasts(xi) < 0) & any(contrasts(xi) > 0))
   }