+++ tests/testthat/test-standardize_parameters.R
@@ -1,8 +1,9 @@
 if (require("testthat") && require("effectsize")) {
+  data("iris")
+  df <- iris
 
-  # "standardize_parameters (simple)" ---------------------------------------
+  # simple ------------------------------------------------------------------
   test_that("standardize_parameters (simple)", {
-    df <- iris
     r <- as.numeric(cor.test(df$Sepal.Length, df$Petal.Length)$estimate)
 
     model <- lm(Sepal.Length ~ Petal.Length, data = df)
@@ -11,33 +12,7 @@ if (require("testthat") && require("effectsize")) {
   })
 
 
-  # standardize_parameters (weighted) -------------------------------
-  test_that("standardize_parameters (weighted)", {
-    mod_w <- lm(mpg ~ disp, data = mtcars, weights = cyl)
-
-    es <- standardize_parameters(mod_w, method = "basic")
-
-    r <- cov.wt(mtcars[, c("disp", "mpg")],
-           wt = mtcars$cyl,
-           cor = TRUE)$cor[1,2]
-
-    testthat::expect_equal(es[2,2], r, tol = 0.001)
-
-
-
-    mod <- lm(mpg ~ disp + factor(am) * hp, data = mtcars)
-    mod_w <- lm(mpg ~ disp + factor(am) * hp, data = mtcars, weights = rep(1,32))
-
-
-    testthat::expect_equal(standardize_info(mod),
-                           standardize_info(mod_w))
-    testthat::expect_equal(standardize_info(mod, robust = TRUE),
-                           standardize_info(mod_w, robust = TRUE),
-                           tol = 0.2)
-
-  })
-
-  # standardize_parameters (model_parameters) -------------------------------
+  # model_parameters -------------------------------
   test_that("standardize_parameters (model_parameters)", {
     model <<- lm(mpg ~ cyl + am, data = mtcars)
     mp <<- parameters::model_parameters(model)
@@ -51,7 +26,7 @@ if (require("testthat") && require("effectsize")) {
     testthat::expect_equal(s1$CI_high, s2$CI_high)
   })
 
-  # "standardize_parameters (lm with ci)" -----------------------------------
+  # lm with ci -----------------------------------
   test_that("standardize_parameters (lm with ci)", {
     model <- lm(Sepal.Length ~ Species + Petal.Width, data = iris)
 
@@ -103,11 +78,18 @@ if (require("testthat") && require("effectsize")) {
       c(0, 0.135, 0.234, 1.073),
       tol = 0.01
     )
+
+
+    m0 <- lm(mpg ~ cyl + factor(am), mtcars)
+    expect_equal(standardize_parameters(m0, method = "refit")[[2]][-1],
+                 standardize_parameters(m0, method = "smart")[[2]][-1], tol = 0.01)
+    expect_equal(standardize_parameters(m0, method = "refit", two_sd = TRUE)[[2]][-1],
+                 standardize_parameters(m0, method = "smart", two_sd = TRUE)[[2]][-1], tol = 0.01)
   })
 
 
-  # "standardize_parameters (with function interactions)" -------------------
-  test_that("standardize_parameters (with function interactions)", {
+  # with function interactions" -------------------
+  test_that("standardize_parameters (with functions /  interactions)", {
     X <- scale(rnorm(100),T,F)
     Z <- scale(rnorm(100),T,F)
     Y <- scale(Z + X * Z + rnorm(100),T,F)
@@ -129,10 +111,27 @@ if (require("testthat") && require("effectsize")) {
     #   standardize_parameters(m1, method = "basic")$Std_Coefficient,
     #   standardize_parameters(m4, method = "basic")$Std_Coefficient
     # )
+
+
+    # transformed resp or pred should not affect
+    mtcars$cyl_exp <- exp(mtcars$cyl)
+    mtcars$mpg_sqrt <- sqrt(mtcars$mpg)
+    m1 <- lm(exp(cyl) ~ am + sqrt(mpg), mtcars)
+    m2 <- lm(cyl_exp ~ am + mpg_sqrt, mtcars)
+
+    expect_message(stdX <- standardize_parameters(m1, method = "refit"))
+    expect_false(isTRUE(all.equal(stdX[[2]],
+                                  standardize_parameters(m2, method = "refit")[[2]])))
+    expect_equal(standardize_parameters(m1, method = "basic")[[2]],
+                 standardize_parameters(m2, method = "basic")[[2]])
+
+    # posthoc / smart don't support data transformation
+    expect_warning(standardize_parameters(m1, method = "smart"))
+    expect_warning(standardize_parameters(m1, method = "posthoc"))
   })
 
 
-  # "standardize_parameters (Bayes)" ----------------------------------------
+  # Bayes ----------------------------------------
   if (require(rstanarm)) {
     test_that("standardize_parameters (Bayes)", {
       testthat::skip_on_cran()
@@ -145,25 +144,25 @@ if (require("testthat") && require("effectsize")) {
       )
 
       testthat::expect_equal(
-        suppressWarnings(standardize_parameters(model, method = "refit")$Std_Median),
+        suppressWarnings(standardize_parameters(model, method = "refit")$Std_Median[1:4]),
         c(0.065, -0.094, -0.100, 0.862),
         tol = 0.01
       )
 
       testthat::expect_equal(
-        suppressWarnings(standardize_parameters(model, method = "posthoc")$Std_Median),
+        suppressWarnings(standardize_parameters(model, method = "posthoc")$Std_Median[1:4]),
         c(0, -0.058, -0.053,  0.838),
         tol = 0.01
       )
 
       posts <- standardize_posteriors(model, method = "posthoc")
-      testthat::expect_equal(dim(posts), c(1000, 4))
+      testthat::expect_equal(dim(posts), c(1000, 5))
       testthat::expect_is(posts, "data.frame")
     })
   }
 
 
-  # "standardize_parameters (Pseudo - GLMM)" --------------------------------
+  # Pseudo - GLMM --------------------------------
   if (require(lme4)) {
     test_that("standardize_parameters (Pseudo - GLMM)", {
       set.seed(1)